requirements.txt
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
Django==3.0.6
django-crispy-forms==1.9.1
djangorestframework==3.11.0
feedparser==5.2.1
kombu==4.6.8
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
vine==1.3.0
